Update Investor views

+++ database/migrations/2026_01_29_164924_create_funds_table.php
@@ -12,10 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('funds', function (Blueprint $table) {
-            $table->id();                       // Local primary key
-            $table->unsignedBigInteger('api_id')->unique(); // API fund ID
-            $table->string('name');             // Fund name
-            $table->timestamps();               // created_at, updated_at
+            $table->id();                       
+            $table->Integer('api_id')->unique(); 
+            $table->string('name');             
+            $table->timestamps();              
         });
     }
